Simplify "different app" logic, rename file, add comment
